# Improve `scan_core` in cupy/core/_routines_math.pyx