This improves `scan_core` ([link](https://github.com/cupy/cupy/blob/e08848f7ed738636ecc3d377fda5f3d69f8a14f9/cupy/core/_routines_math.pyx#L501)) if dtype is promoted.

The current implementation copies the input array before calling a scan kernel when dtype is promoted, but this copying is not necessary when the CuPy bult-in scan kernel is used. This PR eliminates this array copy (when possible) and improves the performance. Below are measured results of `cumsum` on V100.

size | dtype | NumPy | CuPy (current) | CuPy (this PR)
-----|-------|-------|-----|-----
1,000,000 | int8 | 3.401 ms | 0.130 ms | 0.100 ms
1,000,000 | int32 | 2.856 ms | 0.132 ms | 0.095 ms
1,000,000 | uint32 | 3.163 ms | 0.135 ms | 0.095 ms
10,000,000 | int8 | 39.967 ms | 0.816 ms | 0.636 ms
10,000,000 | int32 | 40.706 ms | 0.862 ms | 0.640 ms
10,000,000 | uint32 | 44.313 ms | 0.866 ms | 0.639 ms
100,000,000 | int8 | 390.308 ms | 7.799 ms | 6.179 ms
100,000,000 | int32 | 393.314 ms | 8.273 ms | 6.189 ms
100,000,000 | uint32 | 426.823 ms | 8.275 ms | 6.189 ms

<details>
<summary> Click to see the script </summary>

```python
mport numpy
import cupy
from cupy import testing
from cupyx import time

n_warmup = 1
n_repeat = 10
dtype_set = ('int8', 'int32', 'uint32')
n_set = (1000000, 10000000, 100000000,)


def get_mean_time(perf):
    cpu_time = perf.cpu_times.mean()
    gpu_time = perf.gpu_times.mean()
    return max(cpu_time, gpu_time) * 1000  # ms


def time_cumsum(a):
    cp_a = cupy.array(a)
    ref = numpy.cumsum(a)
    ret = cupy.cumsum(cp_a)
    if a.dtype != numpy.float32:
        testing.assert_allclose(ret, ref)
    times = []
    perf = time.repeat(numpy.cumsum, (a,),
                       n_warmup=n_warmup, n_repeat=n_repeat)
    times.append(get_mean_time(perf))
    perf = time.repeat(cupy.cumsum, (cp_a,),
                       n_warmup=n_warmup, n_repeat=n_repeat)
    times.append(get_mean_time(perf))
    return times


print('size | dtype | NumPy | CuPy')
print('-----|-------|-------|-----')
for n in n_set:
    for dtype in dtype_set:
        a = numpy.random.randint(0, 2, size=n).astype(dtype)
        times = time_cumsum(a)
        print('{} | {} | {:.3f} ms | {:.3f} ms'.format(n, dtype, times[0], times[1]))
```
</details>